Agrego página de mascota encontrada, utilizando un componente comun cambiando el tipo de anuncio
